backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Data Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# File Parsing
pdfplumber==0.10.4
PyPDF2==3.0.1
python-docx==1.1.0

# Database (optional)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# TTS Options
gTTS==2.5.1
edge-tts==6.1.9

# PDF Report Generation
reportlab==4.0.8

# Utilities
python-jose==3.3.0
passlib==1.7.4

# Development
pytest==8.0.0
pytest-asyncio==0.23.4

# Optional: Redis for caching
# redis==5.0.1

# Optional: Celery for background tasks
# celery==5.3.4
